changes the configuration
