+++ src/index.ts
@@ -8,7 +8,10 @@ import {APICache} from './infrastructure/cache/APICache'
 import {securityMiddleware} from './interfaces/http/middleware/security'
 import configTmdb from './infrastructure/config'
 import {setupSwagger} from './docs/swaggerDefinition'
-import {healthApiRouter} from "./interfaces/http/routes/healthApiRouter";
+import {healthApiRouter} from './interfaces/http/routes/healthApiRouter'
+import {addMediaTypeMDB} from './interfaces/http/routes/addMediaType'
+import {favoritesTMDB} from './interfaces/http/routes/favoritesTMDB'
+import {ratedMediaTypeTMDB} from "./interfaces/http/routes/ratedMediaTypeTMDB";
 
 const app = express()
 const PORT = configTmdb.get('PORT')
@@ -20,8 +23,11 @@ app.use(errorHandler())
 
 APICache.initialize(configTmdb.get('CACHE_TIME_OUT'));
 
-app.use('/api/account', routerTMDB)
 app.use('/api', healthApiRouter)
+app.use('/api/account', routerTMDB)
+app.use('/api/account', addMediaTypeMDB)
+app.use('/api/account', favoritesTMDB)
+app.use('/api/account', ratedMediaTypeTMDB)
 
 app.use(errorHandlerApi)
 setupSwagger(app)
